==> talking-data-fraud-detection/01 ParseRawDataAndCreateSampleData.ipynb <==
This notebook deals with creating sample data from train data
1. Read Pickel file
2. Create sample data frame, write to sample_Data directory
3. Start with Features on That Implementation

import pandas as pd
import datetime
import os
import gc

#FILES
TRAIN_FILE_RAW = 'input/train_raw.csv'
TEST_FILE_RAW = 'input/test_raw.csv'

#EDA CONFIG
EDA_CONFIG=True
EDA_NROWS = 1000000
comment_for_file_name="" #'_'+'comment'

def is_train_file(file):
    return 'train' in file
def get_dtypes(file):
    dtypes = {
        'ip'            : 'uint32',
        'app'           : 'uint16',
        'device'        : 'uint16',
        'os'            : 'uint16',
        'channel'       : 'uint16',
        'click_id'      : 'uint32'}
    if is_train_file(file):
        dtypes['is_attributed']='uint8'
    return dtypes
def parsed_file_name(raw_file):
    return raw_file.replace('raw','parsed'+comment_for_file_name)
def parse_raw_file(raw_file):
    parsed_file=None
    if 'raw' in raw_file:
        parsed_file=raw_file.replace('raw','parsed')
    else:
        print(raw_file,' is not a raw file')
        return
    if os.path.isfile(parsed_file):
        os.remove(parsed_file)
        print('deleting old ',parsed_file)
        
    df = pd.read_csv(raw_file,parse_dates=['click_time'])
    print('parsing click_time...')
    df['click_time']=pd.to_datetime(df['click_time']) #convert the click_time 
    print('parsing click_date...')
    df['click_date'] = df['click_time'].dt.date
    print('parsing hour...')
    df['hour'] = pd.to_datetime(df.click_time).dt.hour.astype('uint8')
    print('parsing day')
    df['day'] = pd.to_datetime(df.click_time).dt.day.astype('uint8')
    print('parsing minute')
    df['minute'] = pd.to_datetime(df.click_time).dt.minute.astype('uint8')
    return df

def parse_file(file_name):
    if file_name is None:
        return 'NO FILE NAME'
    print('reading raw file...',file_name)
    df_parsed=parse_raw_file(file_name)
    print('head of file')
    df_parsed.head()
    print('writing file',parsed_file_name(file_name))
    df_parsed.to_csv(parsed_file_name(file_name),index=False)
    return parsed_file_name(file_name),df_parsed

def get_sample_filename(file):
    return file.replace('parsed','sample'+comment_for_file_name)
def get_sample(df_parsed,file):
    if(df_parsed) is None or (file is None):
        return 'GOT NONE, please assign'
    n=None
    if EDA_CONFIG:
        n=EDA_NROWS
        path='eda/'+get_sample_filename(file)
    print('sampling...')
    df_sample=df_parsed.sample(n=EDA_NROWS)
    df_sample_to_write=df_sample.reset_index(drop=True)
    print('writing...')
    df_sample_to_write.to_csv(path,index=False)
    return path,df_sample_to_write

###############
file= TEST_FILE_RAW #TRAIN_FILE_RAW


parsed_file,df_parsed = parse_file(file)
sample_file,df_sample=get_sample(df_parsed,parsed_file)
print('done')

print('done')

df_sample.head()

df_parsed.hour.unique()

df_sample.info()

df_parsed.info()
