==> talking-data-fraud-detection/02 GetTrainValidSamples.ipynb <==
#sample files
#INPUT
TRAIN_SAMPLE='eda/input/train_sample.csv'
#OUTPUT
#train_sample_in eda/input

import pandas as pd

import datetime
TRAIN_HOURS_END = datetime.datetime(2017, 11, 8, 16, 0, 0)
VALID_HOURS_START = datetime.datetime(2017, 11, 9, 4, 0, 0)
VALID_HOURS_END = datetime.datetime(2017, 11, 9, 15, 0, 0)
test_hours=[ 4,  5,  6,  9, 10, 11, 13, 14, 15]
comment='_valid_hr_is_test_hr'
def get_train_valid_from_specific_hours(df):
    df_train_sample= df[df.click_time<=TRAIN_HOURS_END]
    df_valid_sample=df[(df.click_time>=VALID_HOURS_START) & (df.click_time<=VALID_HOURS_END)]
    df_valid_sample=df_valid_sample[df_valid_sample.hour.isin(test_hours)]
    return df_train_sample,df_valid_sample

def get_specific_hours_filepath(file):
    return file.replace('sample','specific_hours_sample'+comment)
def validation_file(file):
    return file.replace('train','valid')

def get_sample(file):
    df=pd.read_csv(file,parse_dates=['click_time','attributed_time','click_date'])
    df_train_sample,df_valid_sample=get_train_valid_from_specific_hours(df)
    df_train_sample.to_csv(get_specific_hours_filepath(file),index=False)
    df_valid_sample.to_csv(get_specific_hours_filepath(validation_file(file)),index=False)
    return get_specific_hours_filepath(file), get_specific_hours_filepath(validation_file(file))

file=TRAIN_SAMPLE
train_sample,valid_sample=get_sample(file)
print('done: ',train_sample,valid_sample)
