==> talking-data-fraud-detection/04.GetSubmission.ipynb <==
c_test_df_prep='test_df_prep.pkl'
c_model='bst.pkl'
predictors = ['app','device','os', 'channel', 'hour','nip_day_test_hh', 'nip_day_hh',
              'nip_hh_os', 'nip_hh_app', 'nip_hh_dev']
categorical = ['app', 'device', 'os', 'channel', 'hour']

import pandas as pd
import lightgbm as lgb

print('load test...')
# test_cols = ['ip','app','device','os', 'channel', 'click_time', 'click_id']
# test_df = pd.read_csv(path+"test.csv", dtype=dtypes, usecols=test_cols)

# test_df = prep_data( test_df )
# gc.collect()
test_df=pd.read_pickle(c_test_df_prep)
sub = pd.DataFrame()
sub['click_id'] = test_df['click_id']

print("Predicting...")
bst = lgb.Booster(model_file=c_model)
sub['is_attributed'] = bst.predict(test_df[predictors])
print("writing...")
sub.to_csv('sub_lgbm_r_to_python.csv',index=False)
print("done...")
print(sub.info())

print('hi!')
