==> talking-data-fraud-detection/05.GetSubmission.ipynb <==
comment='_valid_hr_is_test_hr'
test_features='eda/features/test/all_features.csv'
c_model='eda/models/sample_run/model'+comment
predictors = ['app', 'device', 'os', 'channel', 'hour',
       'day', 'IN_TEST_HH', 'IP_DAY_TEST_HH', 'IP_DAY_HH', 'IP_OS_HH',
       'IP_APP_HH', 'IP_APP_OS_HH', 'APP_DAY_HH']
categorical = ['app', 'device', 'os', 'channel', 'hour','day']

import pandas as pd
import lightgbm as lgb

print('load test...')
test_df=pd.read_csv(test_features)
sub = pd.DataFrame()
sub['click_id'] = test_df['click_id']

print("Predicting...")
bst = lgb.Booster(model_file=c_model)
sub['is_attributed'] = bst.predict(test_df[predictors])
print("writing...")
sub.to_csv('sub_'+comment+'.csv',index=False)
print("done...")
print(sub.info())

print('hi')
