==> talking-data-fraud-detection/09_02_understand PrepData.ipynb <==
import pandas as pd

#read sample pkl file to dataframe
c_train_df_raw_file='train_df_raw_file.pkl'
c_train_df_raw_full_file='train_df_raw_full_file.pkl'
# df_raw_full=pd.read_pickle(c_train_df_raw_full_file)

df=pd.read_pickle(c_train_df_raw_file)

df['hour']=pd.to_datetime(df.click_time).dt.hour.astype('uint8')
df['day']=pd.to_datetime(df.click_time).dt.day.astype('uint8')

df.head()

most_freq_hours_in_test_data = [4, 5, 9, 10, 13, 14]
least_freq_hours_in_test_data = [6, 11, 15]

n=75
df1=df.head(n)

len(df1.ip.unique())


test_df=df1[df1.ip.duplicated(keep=False)]

test_df.head()

test_df.groupby('ip').count()

test_df_2=test_df[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh']).head()

test_df_2.head()

test_df_2.channel.count()

test_df_3=test_df_2[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].count().reset_index().rename(index=str, 
             columns={'channel': 'nip_day_test_hh'})

test_df_2.head()

test_df_3.head()

test_df_2.merge(test_df_3,on=['ip','day'],how='left')

test_df_2.merge(test_df_3,on=['ip','day','in_test_hh'],how='left')

test_df_2.merge(test_df_3,on=['ip','day','in_test_hh'],how='left')

test_df_2.merge(test_df_3,on=['ip'],how='right')

df1[['ip', 'day', 'in_test_hh', 'channel']].head(n)

df1[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].head()

df1[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].count().head()


df1[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].count().reset_index().rename(index=str, 
             columns={'channel': 'nip_day_test_hh'})

df2

df2.groupby('X')['Y'].count()

df2.groupby(['X'],sort=True).count().head(20)

df['in_test_hh']=(3-2*df.hour.isin(most_freq_hours_in_test_data)-df.hour.isin(least_freq_hours_in_test_data))

df[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].head()

df[['ip', 'day', 'in_test_hh', 'channel']].groupby(by=['ip', 'day',
             'in_test_hh'])[['channel']].count().reset_index().rename(index=str, 
             columns={'channel': 'nip_day_test_hh'})

df_raw.head()

def prep_data(df):
    df['hour'] = pd.to_datetime(df.click_time).dt.hour.astype('uint8')
    df['day'] = pd.to_datetime(df.click_time).dt.day.astype('uint8')
    return df


df=prep_data(df_raw)

df.head()

df.hour.unique()
