==> talking-data-fraud-detection/09_04.ValidationSet.ipynb <==
The rationale behind getting the validation set is that it(train,validation) should have the same distribution as (train,test)
For the current problem we have training from 6-16:00 to 9-1500 in train and 10-4:00 to 10-15 in test sets.


So to have a match of the LB and a valid Validation set, we can take train as , truncate train from 6-16:00 to 

 train date ranges
# min: 06 14:32:21 max: 09 16:00:00

# test date ranges
# min: 10 04:00:00 max: 10 15:00:00

train-validation SPLIT
# x0: 2017-11-06 14:32:21 -> 2017-11-08 16:00:00
# x1: 2017-11-09 04:00:00 -> 2017-11-09 15:00:00   

import pandas as pd


path='input/'
train='train.csv'

df=pd.read_csv(path+train,parse_dates=['click_time','attributed_time'])

#identify the dataframe that has the following dates
df.
